refactor(firmware): Split RmcsBoard app into a superbuild layout
- Turn firmware/rmcs_board into a superbuild that delegates to app/
- Move the existing rmcs_board application sources under firmware/rmcs_board/app
- Update CI, lint targets, and firmware packaging for the new output layout
- Prepare rmcs_board for a future bootloader subbuild
