+++ firmware/rmcs_board/app/src/spi/spi.hpp
@@ -16,7 +16,7 @@
 
 #include "core/src/utility/assert.hpp"
 #include "core/src/utility/immovable.hpp"
-#include "firmware/rmcs_board/src/utility/lazy.hpp"
+#include "firmware/rmcs_board/app/src/utility/lazy.hpp"
 
 namespace librmcs::firmware::spi {
