+++ firmware/rmcs_board/app/src/uart/tx_buffer.hpp
@@ -21,7 +21,7 @@
 
 #include "core/include/librmcs/data/datas.hpp"
 #include "core/src/utility/assert.hpp"
-#include "firmware/rmcs_board/src/utility/ring_buffer.hpp"
+#include "firmware/rmcs_board/app/src/utility/ring_buffer.hpp"
 
 namespace librmcs::firmware::uart {
