+++ firmware/rmcs_board/app/src/utility/assert.cpp
